feat(iac): Refactor comparison of template dictionary to actual configuration to allow for extra keys in the actual configuration [pulumi up]
This change is important because preventing the addition of new keys can hinder development. If new configuration keys are required but not yet present in the environment secrets or the stack config, deployments will break when merging the code that introduces them.

Additionally, if a developer tries to add the keys before merging, it may block other code from being merged until their changes are integrated. This creates unnecessary pressure on developers.